+++ ChangeLog.md
@@ -5,6 +5,7 @@
 - DicomElement.ValueRepresentation.ValidateString() now throws a DicomValidationException if a null value is passed (#1590)
 - The optional AffectedSopInstanceUID is added to the CStoreResponse by default. (#1390)
 - Handle invalid DICOM files, that contain "," as decimal separator in DS values (#1296)
+- DicomDataset.AddOrUpdate also allows passing DicomSequence as argument type (#1664)
 
 ### 5.1.5 (2024-11-25)